Azure AI Services - Web with Azure AI Services/requirements.txt
streamlit
azure-cognitiveservices-speech
streamlit-extras